HTML CSS JAVASCRIPT
This coding project  helps  to limit the number of words in articles to make them look more eye-catching and attractive. This also helps to keep us updated with the present number of words while typing which gives us an idea about when to wrap up. This updates the count of words in real-time in text space. We can also upgrade it to count the number of lines, paragraphs, or characters. The prerequisites are HTML, CSS, and modern frameworks of JavaScript( for a better User Interface).